Allow query parameters
app.py:

 * imports
Add request, redirect, url_for from flask library to support query parameter usage and redirecting from the root page to the license page.

 * /
Change behavior of root to redirect to license page.

 * /license
Add get for parameter, then check if the parameter was passed and load the page if so. If not, load the default HTML.

 * /badge
Same changes as license.

 * /name
Same changes as license.

This change is adding the ability for users to use query parameters if they want. Next commit will be to update the HTML to utilize these new changes.