Clean up data files
Clean up existing CSV file

Create 1-312-data sqlite DB file. It's a binary so not ideal to store in git but doing this so it can be pulled into the Glitch app without having to waste cycles reading the CSV into the database. We will leave the CSV for now.