Applying comments, mostly
app.py:
 * Reorder entities so search fields appear as
    - First Name
    - Last Name
    - Badge Number

* Revert query param gets back to just simple `.get`s to work with new query building format.

dataset.py:
 * Update typing import to match reality.
 * Add table to explain database format
 * Add RosterRecord class to simplify working with database records
 * Update _sort_names to make use of new RosterRecord class
 * Broke up raw HTML string to conform to PEP8 or which ever the "no long lines" one is.
 * Update _augment_with_salary to make use of new RosterRecord Class
 * Update name_lookup function a couple ways which I will explain below

name_lookup was updated so that the order of the inputs for the function followed a more natural (to me at least) ordering. First, Last, Badge#. I've also moved the empty return to the top as suggested.

I tried simplifying the code a little bit so we aren't relying on a bunch of if statements to build queries. Let me know what you think on that _build_sql_query function. It could probably be simplified further, not sure.